// src/api/generated/model/getApiV1OrganizationsOrgIdFinancialMovementsParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Nummo API
 * API de administración financiera y cartera (V1): auth multi-tenant, contactos, cartera con mora, pagos, gastos/egresos, caja y reportes. Autenticación por cookie de sesión (HttpOnly) + CSRF; dinero como string decimal.
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiV1OrganizationsOrgIdFinancialMovementsDirection } from './getApiV1OrganizationsOrgIdFinancialMovementsDirection';
import type { GetApiV1OrganizationsOrgIdFinancialMovementsMovementType } from './getApiV1OrganizationsOrgIdFinancialMovementsMovementType';
import type { GetApiV1OrganizationsOrgIdFinancialMovementsOrder } from './getApiV1OrganizationsOrgIdFinancialMovementsOrder';

export type GetApiV1OrganizationsOrgIdFinancialMovementsParams = {
/**
 * @minimum 1
 * @maximum 9007199254740991
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: number;
/**
 * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
 */
financialAccountId?: string;
direction?: GetApiV1OrganizationsOrgIdFinancialMovementsDirection;
movementType?: GetApiV1OrganizationsOrgIdFinancialMovementsMovementType;
/**
 * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
 */
from?: string;
/**
 * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
 */
to?: string;
order?: GetApiV1OrganizationsOrgIdFinancialMovementsOrder;
};